// src/types/trade.ts
export enum TradeSymbol {
  PreciousStones = "PRECIOUS_STONES",
  QuartzSand = "QUARTZ_SAND",
  SiliconCrystals = "SILICON_CRYSTALS",
  AmmoniaIce = "AMMONIA_ICE",
  LiquidHydrogen = "LIQUID_HYDROGEN",
  LiquidNitrogen = "LIQUID_NITROGEN",
  IceWater = "ICE_WATER",
  ExoticMatter = "EXOTIC_MATTER",
  AdvancedCircuitry = "ADVANCED_CIRCUITRY",
  GravitonEmitters = "GRAVITON_EMITTERS",
  Iron = "IRON",
  IronOre = "IRON_ORE",
  Copper = "COPPER",
  CopperOre = "COPPER_ORE",
  Aluminum = "ALUMINUM",
  AluminumOre = "ALUMINUM_ORE",
  Silver = "SILVER",
  SilverOre = "SILVER_ORE",
  Gold = "GOLD",
  GoldOre = "GOLD_ORE",
  Platinum = "PLATINUM",
  PlatinumOre = "PLATINUM_ORE",
  Diamonds = "DIAMONDS",
  Uranite = "URANITE",
  UraniteOre = "URANITE_ORE",
  Meritium = "MERITIUM",
  MeritiumOre = "MERITIUM_ORE",
  Hydrocarbon = "HYDROCARBON",
  Antimatter = "ANTIMATTER",
  Fertilizers = "FERTILIZERS",
  Fabrics = "FABRICS",
  Food = "FOOD",
  Jewelry = "JEWELRY",
  Machinery = "MACHINERY",
  Firearms = "FIREARMS",
  AssaultRifles = "ASSAULT_RIFLES",
  MilitaryEquipment = "MILITARY_EQUIPMENT",
  Explosives = "EXPLOSIVES",
  LabInstruments = "LAB_INSTRUMENTS",
  Ammunition = "AMMUNITION",
  Electronics = "ELECTRONICS",
  ShipPlating = "SHIP_PLATING",
  Equipment = "EQUIPMENT",
  Fuel = "FUEL",
  Medicine = "MEDICINE",
  Drugs = "DRUGS",
  Clothing = "CLOTHING",
  Microprocessors = "MICROPROCESSORS",
  Plastics = "PLASTICS",
  Polynucleotides = "POLYNUCLEOTIDES",
  Biocomposites = "BIOCOMPOSITES",
  Nanobots = "NANOBOTS",
  AiMainframes = "AI_MAINFRAMES",
  QuantumDrives = "QUANTUM_DRIVES",
  RoboticDrones = "ROBOTIC_DRONES",
  CyberImplants = "CYBER_IMPLANTS",
  GeneTherapeutics = "GENE_THERAPEUTICS",
  NeuralChips = "NEURAL_CHIPS",
  MoodRegulators = "MOOD_REGULATORS",
  ViralAgents = "VIRAL_AGENTS",
  MicroFusionGenerators = "MICRO_FUSION_GENERATORS",
  Supergrains = "SUPERGRAINS",
  LaserRifles = "LASER_RIFLES",
  Holographics = "HOLOGRAPHICS",
  ShipSalvage = "SHIP_SALVAGE",
  RelicTech = "RELIC_TECH",
  NovelLifeforms = "NOVEL_LIFEFORMS",
  BotanicalSpecimens = "BOTANICAL_SPECIMENS",
  CulturalArtifacts = "CULTURAL_ARTIFACTS",
  ReactorSolarI = "REACTOR_SOLAR_I",
  ReactorFusionI = "REACTOR_FUSION_I",
  ReactorFissionI = "REACTOR_FISSION_I",
  ReactorChemicalI = "REACTOR_CHEMICAL_I",
  ReactorAntimatterI = "REACTOR_ANTIMATTER_I",
  EngineImpulseDriveI = "ENGINE_IMPULSE_DRIVE_I",
  EngineIonDriveI = "ENGINE_ION_DRIVE_I",
  EngineIonDriveII = "ENGINE_ION_DRIVE_II",
  EngineHyperDriveI = "ENGINE_HYPER_DRIVE_I",
  ModuleMineralProcessorI = "MODULE_MINERAL_PROCESSOR_I",
  ModuleCargoHoldI = "MODULE_CARGO_HOLD_I",
  ModuleCrewQuartersI = "MODULE_CREW_QUARTERS_I",
  ModuleEnvoyQuartersI = "MODULE_ENVOY_QUARTERS_I",
  ModulePassengerCabinI = "MODULE_PASSENGER_CABIN_I",
  ModuleMicroRefineryI = "MODULE_MICRO_REFINERY_I",
  ModuleOreRefineryI = "MODULE_ORE_REFINERY_I",
  ModuleFuelRefineryI = "MODULE_FUEL_REFINERY_I",
  ModuleScienceLabI = "MODULE_SCIENCE_LAB_I",
  ModuleJumpDriveI = "MODULE_JUMP_DRIVE_I",
  ModuleJumpDriveII = "MODULE_JUMP_DRIVE_II",
  ModuleJumpDriveIII = "MODULE_JUMP_DRIVE_III",
  ModuleWarpDriveI = "MODULE_WARP_DRIVE_I",
  ModuleWarpDriveII = "MODULE_WARP_DRIVE_II",
  ModuleWarpDriveIII = "MODULE_WARP_DRIVE_III",
  ModuleShieldGeneratorI = "MODULE_SHIELD_GENERATOR_I",
  ModuleShieldGeneratorII = "MODULE_SHIELD_GENERATOR_II",
  MountGasSiphonI = "MOUNT_GAS_SIPHON_I",
  MountGasSiphonII = "MOUNT_GAS_SIPHON_II",
  MountGasSiphonIII = "MOUNT_GAS_SIPHON_III",
  MountSurveyorI = "MOUNT_SURVEYOR_I",
  MountSurveyorII = "MOUNT_SURVEYOR_II",
  MountSurveyorIII = "MOUNT_SURVEYOR_III",
  MountSensorArrayI = "MOUNT_SENSOR_ARRAY_I",
  MountSensorArrayII = "MOUNT_SENSOR_ARRAY_II",
  MountSensorArrayIII = "MOUNT_SENSOR_ARRAY_III",
  MountMiningLaserI = "MOUNT_MINING_LASER_I",
  MountMiningLaserII = "MOUNT_MINING_LASER_II",
  MountMiningLaserIII = "MOUNT_MINING_LASER_III",
  MountLaserCannonI = "MOUNT_LASER_CANNON_I",
  MountMissileLauncherI = "MOUNT_MISSILE_LAUNCHER_I",
  MountTurretI = "MOUNT_TURRET_I"
}